[MODORDSTOR-298] - Refactoring
